chore(Mapper): rename `setRecursively` to `setByPath`
